dbx-tools/pyproject.toml
[project]
name = "dbx-tools"
version = "0.0.1+g3f310a4"
requires-python = ">=3.12,<3.13"
dependencies = [
    "databricks-connect>=17.0,<17.3",
    "databricks-sdk>=0.73.0",
    "pydantic",
    "reggie-core @ file://${PROJECT_ROOT}/../reggie-core",
]
[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources.reggie-core]
workspace = true
